Use size_t for stuff that counts a number of items
This is unsigned, so it fixes a number of signed vs. unsigned comparison warnings.

Contributes to issue CURA-4410.

+++ src/commandSocket.cpp
@@ -54,7 +54,7 @@ public:
     Arcus::Socket* socket;
     
     // Number of objects that need to be sliced
-    int object_count;
+    size_t object_count;
 
     std::string temp_gcode_file;
     std::ostringstream gcode_output_stream;
